+++ ts/.prettierignore
@@ -13,11 +13,7 @@ pnpm-lock.yaml
 # Typescript config
 tsconfig.json
 
-# TypeChat Fork
-packages/typechat
-
-
 # Test data
 packages/defaultAgentProvider/test/data
 packages/defaultAgentProvider/test/repo
-packages/dispatcher/data/**/*.json
+packages/defaultAgentProvider/data/**/*.json
