+++ ts/packages/cli/src/commands/interactive.ts
@@ -2,28 +2,35 @@
 // Licensed under the MIT License.
 
 import { Args, Command, Flags } from "@oclif/core";
-import { getCacheFactory } from "agent-dispatcher/internal";
+import { createDispatcher, Dispatcher } from "agent-dispatcher";
+import {
+    createActionConfigProvider,
+    getCacheFactory,
+    getInstanceDir,
+    getSchemaNamesForActionConfigProvider,
+} from "agent-dispatcher/internal";
 import {
     getDefaultAppAgentProviders,
     getDefaultConstructionProvider,
 } from "default-agent-provider";
-import { getSchemaNamesFromDefaultAppAgentProviders } from "default-agent-provider/internal";
 import inspector from "node:inspector";
 import { getChatModelNames } from "aiclient";
 import {
     processCommands,
     withConsoleClientIO,
 } from "agent-dispatcher/helpers/console";
-import { createDispatcher, Dispatcher } from "agent-dispatcher";
 
 const modelNames = await getChatModelNames();
-
+const defaultAppAgentProviders = getDefaultAppAgentProviders(getInstanceDir());
+const schemaNames = getSchemaNamesForActionConfigProvider(
+    await createActionConfigProvider(defaultAppAgentProviders),
+);
 export default class Interactive extends Command {
     static description = "Interactive mode";
     static flags = {
         translator: Flags.string({
             description: "Schema names",
-            options: getSchemaNamesFromDefaultAppAgentProviders(),
+            options: schemaNames,
             multiple: true,
         }),
         explainer: Flags.string({
@@ -69,7 +76,7 @@ export default class Interactive extends Command {
 
         await withConsoleClientIO(async (clientIO) => {
             const dispatcher = await createDispatcher("cli interactive", {
-                appAgentProviders: getDefaultAppAgentProviders(),
+                appAgentProviders: defaultAppAgentProviders,
                 schemas,
                 translation: { model: flags.model },
                 explainer: { name: flags.explainer },
