--- ts/packages/defaultAgentProvider/src/internal.ts
@@ -1,9 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-export {
-    getSchemaNamesFromDefaultAppAgentProviders,
-    getActionConfigProviderFromDefaultAppAgentProviders,
-    createSchemaInfoProviderFromDefaultAppAgentProviders,
-} from "./defaultAgentProviders.js";
-export { getTestDataFiles } from "./utils/config.js";
