+++ ts/packages/defaultAgentProvider/test/construction.spec.ts
@@ -4,9 +4,14 @@
 import dotenv from "dotenv";
 dotenv.config({ path: new URL("../../../../.env", import.meta.url) });
 
-import { getCacheFactory, readTestData } from "agent-dispatcher/internal";
+import {
+    createSchemaInfoProvider,
+    getCacheFactory,
+    readTestData,
+    createActionConfigProvider,
+} from "agent-dispatcher/internal";
 import { Actions, RequestAction } from "agent-cache";
-import { createSchemaInfoProviderFromDefaultAppAgentProviders } from "../src/defaultAgentProviders.js";
+import { getDefaultAppAgentProviders } from "../src/defaultAgentProviders.js";
 import { glob } from "glob";
 
 const dataFiles = ["test/data/**/v5/*.json"];
@@ -30,6 +35,9 @@ const matchConfig = {
     rejectReferences: false,
 };
 
+const schemaInfoProvider = createSchemaInfoProvider(
+    await createActionConfigProvider(getDefaultAppAgentProviders(undefined)),
+);
 describe("construction", () => {
     describe("roundtrip", () => {
         it.each(testInput)(
@@ -50,8 +58,7 @@ describe("construction", () => {
                     requestAction,
                     explanation,
                     {
-                        schemaInfoProvider:
-                            createSchemaInfoProviderFromDefaultAppAgentProviders(),
+                        schemaInfoProvider,
                     },
                 );
