+++ ts/packages/defaultAgentProvider/test/translate.test.ts
@@ -9,8 +9,9 @@ import {
     readTestData,
     loadAgentJsonTranslator,
 } from "agent-dispatcher/internal";
+import { createActionConfigProvider } from "agent-dispatcher/internal";
 import { JSONAction } from "agent-cache";
-import { getActionConfigProviderFromDefaultAppAgentProviders } from "../src/defaultAgentProviders.js";
+import { getDefaultAppAgentProviders } from "../src/defaultAgentProviders.js";
 const dataFiles = [
     "test/data/player/v5/simple.json",
     "test/data/player/v5/full.json",
@@ -29,13 +30,16 @@ const testInput = inputs.flatMap((f) =>
     ]),
 );
 
+const provider = await createActionConfigProvider(
+    getDefaultAppAgentProviders(undefined),
+);
 describe("translation", () => {
     it.each(testInput)(
         "translate [%s] '%s'",
         async (translatorName, request, action) => {
             const translator = loadAgentJsonTranslator(
                 translatorName,
-                getActionConfigProviderFromDefaultAppAgentProviders(),
+                provider,
             );
             const result = await translator.translate(request);
             expect(result.success).toBe(true);
