+++ ts/packages/dispatcher/src/npmAgentProvider.ts
@@ -1,0 +1,7 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export {
+    AppAgentInfo,
+    createNpmAppAgentProvider,
+} from "./agentProvider/npmAgentProvider.js";
