+++ ts/packages/dispatcher/src/translation/agentTranslators.ts
@@ -6,12 +6,7 @@ import {
     createJsonTranslatorFromSchemaDef,
     enableJsonTranslatorStreaming,
 } from "common-utils";
-import {
-    AppAction,
-    ActionManifest,
-    SchemaDefinition,
-    AppAgentManifest,
-} from "@typeagent/agent-sdk";
+import { AppAction } from "@typeagent/agent-sdk";
 import { Result, TypeChatJsonTranslator } from "typechat";
 import { getPackageFilePath } from "../utils/getPackageFilePath.js";
 import { getMultipleActionSchemaDef } from "./multipleActionSchema.js";
@@ -21,7 +16,6 @@ import {
     IncrementalJsonValueCallBack,
 } from "common-utils";
 
-import registerDebug from "debug";
 import { HistoryContext, ParamObjectType } from "agent-cache";
 import { createTypeAgentRequestPrompt } from "../context/chatHistoryPrompt.js";
 import {
@@ -35,104 +29,9 @@ import {
     generateSchemaTypeDefinition,
     ActionSchemaObject,
     ActionSchemaCreator as sc,
-    ActionSchemaFile,
 } from "action-schema";
-const debugConfig = registerDebug("typeagent:dispatcher:schema:config");
-
-// A flatten AppAgentManifest
-export type ActionConfig = {
-    emojiChar: string;
-
-    schemaDefaultEnabled: boolean;
-    actionDefaultEnabled: boolean;
-    transient: boolean;
-    schemaName: string;
-} & SchemaDefinition;
-
-export interface ActionConfigProvider {
-    tryGetActionConfig(schemaName: string): ActionConfig | undefined;
-    getActionConfig(schemaName: string): ActionConfig;
-    getActionConfigs(): [string, ActionConfig][];
-    getActionSchemaFileForConfig(config: ActionConfig): ActionSchemaFile;
-}
-
-function isValidSubSchemaName(schemaNamePart: string) {
-    // . is use as a sub-schema separator
-    // | is used in the cache as as multiple schema name separator
-    // , is used in the cache as a separator between schema name and its hash
-    return !/[.|,]/.test(schemaNamePart);
-}
-
-function collectActionConfigs(
-    actionConfigs: { [key: string]: ActionConfig },
-    manifest: ActionManifest,
-    schemaName: string,
-    emojiChar: string,
-    transient: boolean,
-    schemaDefaultEnabled: boolean,
-    actionDefaultEnabled: boolean,
-) {
-    transient = manifest.transient ?? transient; // inherit from parent if not specified
-    schemaDefaultEnabled =
-        manifest.schemaDefaultEnabled ??
-        manifest.defaultEnabled ??
-        schemaDefaultEnabled; // inherit from parent if not specified
-    actionDefaultEnabled =
-        manifest.actionDefaultEnabled ??
-        manifest.defaultEnabled ??
-        actionDefaultEnabled; // inherit from parent if not specified
-
-    if (manifest.schema) {
-        debugConfig(`Adding schema '${schemaName}'`);
-        actionConfigs[schemaName] = {
-            schemaName,
-            emojiChar,
-            ...manifest.schema,
-            transient,
-            schemaDefaultEnabled,
-            actionDefaultEnabled,
-        };
-    }
-
-    const subManifests = manifest.subActionManifests;
-    if (subManifests) {
-        for (const [subName, subManfiest] of Object.entries(subManifests)) {
-            if (!isValidSubSchemaName(subName)) {
-                throw new Error(`Invalid sub-schema name: ${subName}`);
-            }
-            collectActionConfigs(
-                actionConfigs,
-                subManfiest,
-                `${schemaName}.${subName}`,
-                emojiChar,
-                transient, // propagate default transient
-                schemaDefaultEnabled, // propagate default schemaDefaultEnabled
-                actionDefaultEnabled, // propagate default actionDefaultEnabled
-            );
-        }
-    }
-}
-
-export function convertToActionConfig(
-    name: string,
-    config: AppAgentManifest,
-    actionConfigs: Record<string, ActionConfig> = {},
-): Record<string, ActionConfig> {
-    if (!isValidSubSchemaName(name)) {
-        throw new Error(`Invalid schema name: ${name}`);
-    }
-    const emojiChar = config.emojiChar;
-    collectActionConfigs(
-        actionConfigs,
-        config,
-        name,
-        emojiChar,
-        false, // transient default to false if not specified
-        true, // translationDefaultEnable default to true if not specified
-        true, // actionDefaultEnabled default to true if not specified
-    );
-    return actionConfigs;
-}
+import { ActionConfig } from "./actionConfig.js";
+import { ActionConfigProvider } from "./actionConfigProvider.js";
 
 export function getAppAgentName(schemaName: string) {
     return schemaName.split(".")[0];
